django_project/requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.5.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.8
natsort==8.4.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==2.0.4
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.18.2
